date 18.04.2025 time 17 houre 05 minute
in file ExpressionMaker.cpp:1685:81:
fixed error

error: call of overloaded 'CharString(long int)' is ambiguous
CharString val(strtol(literalLxm.GetBuf().c_str(), NULL, base) );
                                                                                 ^
In file ExpressionMaker.cpp:4:
nrc.h:101:17: note:
candidate: 'nrc::CharString::CharString(double, nrc::SObject*)'
CharString( double f, SObject *aobj = NULL ) {

nrc.h:95:17: note:
candidate: 'nrc::CharString::CharString(int, nrc::SObject*)'
CharString( int d, SObject *aobj = NULL ) {

nrc.h:80:15: note:
candidate: 'nrc::CharString::CharString(const nrc::CharString&)'
class CharString : public SObject, public string

+++ nrcpp/ExpressionMaker.cpp
@@ -1682,7 +1682,7 @@ POperand LiteralMaker::Make()
 		int unsign = lc == INTEGER16 || lc == UINTEGER8 ? KWUNSIGNED : -1;
 		int base = lc == INTEGER16 || lc == UINTEGER16 ? 16 : 8;
 
-		CharString val( strtol(literalLxm.GetBuf().c_str(), NULL, base) );
+		CharString val((int) strtol(literalLxm.GetBuf().c_str(), NULL, base) );
 		literal = new Literal( 
 			const_cast<BaseType*>(
 			&ImplicitTypeManager(KWINT, unsign).GetImplicitType()), true, false, dtl, val );
